Refactor project into standalone scripts
- Moved all files from `src/display_jira_tickets/` directly to `src/`.
- Renamed `__main__.py` to `display.py` and added executable shebang.
- Modified imports to absolute imports matching standard script execution.
- Updated `pyproject.toml` to list modules instead of searching for a package.
- Adjusted tests and documentation (README, AGENTS.md) to reflect the new structure.
